+++ CHANGELOG.md
@@ -5,12 +5,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Upgraded
-- "Keyple Util Library" to version `2.1.0` by removing the use of deprecated methods.
+### Added
+- `Plugin.getReaderExtension` to access the reader's extension class.
+### Deprecated
+- `Reader` in favor of `CardReader` from the "Calypsonet Terminal Reader API".
+- `ObservableReader` in favor of `ObservableCardReader` from the "Calypsonet Terminal Reader API".
+- `ConfigurableReader` in favor of `ConfigurableCardReader` from the "Calypsonet Terminal Reader API".
+- `ReaderEvent` in favor of `CardReaderEvent` from the "Calypsonet Terminal Reader API".
 ### Fixed
 - Auto management of 61XX and 6CXX status words and case 4 commands (Calypsonet Terminal Requirements: `RL-SW-61XX.1`, `RL-SW-6CXX.1`, `RL-SW-ANALYSIS.1` and `RL-SW-CASE4.1`) (issue [#37]).
 - Returned value of `getActiveSmartCard` method when there is no active smart card (issue [#40]).
 - JSON serialization for interfaces in objects trees (issue [#43]).
+### Upgraded
+- "Keyple Util Library" to version `2.1.0` by removing the use of deprecated methods.
 
 ## [2.0.1] - 2021-12-08
 ### Added